op-bindings: separate bindings into new module to resolve dependency problems
