+++ op-bindings/doc.go
@@ -7,4 +7,4 @@ To regenerate the bindings, run `make`
 The following programs are required: `jq`, `abigen`, and `solc` (at version 0.8.10)
 */
 
-package contracts
+package op_bindings
