+++ op-bindings/l2oo/l2_output_oracle.go
