+++ op-e2e/setup.go
@@ -10,15 +10,15 @@ import (
 	"time"
 
 	bss "github.com/ethereum-optimism/optimism/op-batcher"
-	"github.com/ethereum-optimism/optimism/op-node/contracts/deposit"
-	"github.com/ethereum-optimism/optimism/op-node/contracts/l1block"
-	"github.com/ethereum-optimism/optimism/op-node/contracts/withdrawer"
+	"github.com/ethereum-optimism/optimism/op-bindings/deposit"
+	"github.com/ethereum-optimism/optimism/op-bindings/l1block"
+	"github.com/ethereum-optimism/optimism/op-bindings/l2oo"
+	"github.com/ethereum-optimism/optimism/op-bindings/withdrawer"
 	rollupNode "github.com/ethereum-optimism/optimism/op-node/node"
 	"github.com/ethereum-optimism/optimism/op-node/p2p"
 	"github.com/ethereum-optimism/optimism/op-node/predeploy"
 	"github.com/ethereum-optimism/optimism/op-node/rollup"
 	l2os "github.com/ethereum-optimism/optimism/op-proposer"
-	"github.com/ethereum-optimism/optimism/op-proposer/bindings/l2oo"
 
 	"github.com/ethereum/go-ethereum/accounts"
 	"github.com/ethereum/go-ethereum/accounts/abi/bind"
