+++ op-node/go.mod
@@ -3,7 +3,8 @@ module github.com/ethereum-optimism/optimism/op-node
 go 1.18
 
 require (
-	github.com/ethereum/go-ethereum v1.10.16
+	github.com/ethereum-optimism/optimism/op-bindings v0.0.0
+	github.com/ethereum/go-ethereum v1.10.17
 	github.com/golang/snappy v0.0.4
 	github.com/google/go-cmp v0.5.8
 	github.com/hashicorp/go-multierror v1.0.0
@@ -171,7 +172,9 @@ require (
 	gopkg.in/yaml.v3 v3.0.0-20210107192922-496545a6307b // indirect
 )
 
-replace github.com/ethereum/go-ethereum v1.10.16 => github.com/ethereum-optimism/reference-optimistic-geth v0.0.0-20220428214415-643b449b5818
+replace github.com/ethereum-optimism/optimism/op-bindings v0.0.0 => ../op-bindings
+
+replace github.com/ethereum/go-ethereum v1.10.17 => github.com/ethereum-optimism/reference-optimistic-geth v0.0.0-20220428214415-643b449b5818
 
 // For local debugging:
-// replace github.com/ethereum/go-ethereum v1.10.16 => ../go-ethereum
+// replace github.com/ethereum/go-ethereum v1.10.17 => ../go-ethereum
