+++ op-proposer/go.mod
@@ -3,8 +3,9 @@ module github.com/ethereum-optimism/optimism/op-proposer
 go 1.18
 
 require (
+	github.com/ethereum-optimism/optimism/op-bindings v0.0.0
 	github.com/ethereum-optimism/optimism/op-node v0.0.0
-	github.com/ethereum/go-ethereum v1.10.16
+	github.com/ethereum/go-ethereum v1.10.17
 	github.com/miguelmota/go-ethereum-hdwallet v0.1.1
 	github.com/stretchr/testify v1.7.0
 	github.com/urfave/cli v1.22.5
@@ -154,4 +155,6 @@ require (
 
 replace github.com/ethereum-optimism/optimism/op-node v0.0.0 => ../op-node
 
-replace github.com/ethereum/go-ethereum v1.10.16 => github.com/ethereum-optimism/reference-optimistic-geth v0.0.0-20220427171107-d0070e0a6ead
+replace github.com/ethereum-optimism/optimism/op-bindings v0.0.0 => ../op-bindings
+
+replace github.com/ethereum/go-ethereum v1.10.17 => github.com/ethereum-optimism/reference-optimistic-geth v0.0.0-20220427171107-d0070e0a6ead
